+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2026-01-03
+
+### Added
+
+- `sync` feature to compile blocking driver implementation
+- Rp2040 async examples in `examples/rp2040_async` directory
+
+### Changed
+
+- Driver implementation is now async by default
+- Tests are now async by default
+
 ## [1.0.0] - 2025-09-29
 
 ### Added